+++ src/main/java/chess/domain/game/ChessGame.java
@@ -4,7 +4,7 @@ import chess.domain.Position;
 import chess.domain.Score;
 import chess.domain.TeamColor;
 import chess.domain.piece.*;
-import chess.view.PieceDto;
+import chess.controller.dto.PieceDto;
 
 import java.util.List;
 import java.util.Set;
